+++ examples/matrices.js
@@ -50,7 +50,7 @@ const createMatrixPiles = async element => {
     previewAggregator: matrixPreviewAggregator,
     items: data,
     itemSize: 64,
-    pileCellAlign: 'center',
+    pileCellAlignment: 'center',
     pileScale: pile => 1 + Math.min((pile.items.length - 1) * 0.05, 0.5)
   });