+++ src/library.js
@@ -193,6 +193,8 @@ const createPilingJs = (rootElement, initOptions = {}) => {
       }
     },
     previewBackgroundOpacity: true,
+    randomOffsetRange: true,
+    randomRotationRange: true,
     renderer: {
       get: 'itemRenderer',
       set: value => [createAction.setItemRenderer(value)]
@@ -961,19 +963,23 @@ const createPilingJs = (rootElement, initOptions = {}) => {
     const movingPiles = [];
     items.forEach((itemId, index) => {
       const pile = pileInstances.get(itemId);
+      const item = renderedItems.get(itemId);
       const tweener = createTweener({
         duration: 250,
         delay: 0,
         interpolator: interpolateVector,
-        endValue:
-          itemPositions.length > 0
+        endValue: [
+          ...(itemPositions.length > 0
             ? itemPositions[index]
-            : renderedItems.get(itemId).originalPosition,
+            : item.originalPosition),
+          0
+        ],
         getter: () => {
-          return [pile.x, pile.y];
+          return [pile.x, pile.y, item.sprite.angle];
         },
-        setter: xy => {
-          pile.moveTo(...xy);
+        setter: newValue => {
+          pile.moveTo(newValue[0], newValue[1]);
+          item.sprite.angle = newValue[2];
         },
         onDone: finalValue => {
           movingPiles.push({
@@ -1538,16 +1544,24 @@ const createPilingJs = (rootElement, initOptions = {}) => {
       renderRaf();
     }
 
-    if (state.scaledPile !== newState.scaledPile) {
-      if (state.scaledPile.length !== 0) {
-        if (pileInstances.has(state.scaledPile[0])) {
-          const pile = pileInstances.get(state.scaledPile[0]).graphics;
-          pile.scale(1);
-          updateBoundingBox(state.scaledPile[0]);
+    if (state.scaledPiles !== newState.scaledPiles) {
+      state.scaledPiles
+        .map(scaledPile => pileInstances.get(scaledPile))
+        .filter(scaledPileInstance => scaledPileInstance)
+        .forEach(scaledPileInstance => {
+          scaledPileInstance.scale(1);
+          updateBoundingBox(scaledPileInstance.id);
           activePile.removeChildren();
-          normalPiles.addChild(pile);
-        }
-      }
+          normalPiles.addChild(scaledPileInstance.graphics);
+        });
+
+      newState.scaledPiles
+        .map(scaledPile => pileInstances.get(scaledPile))
+        .filter(scaledPileInstance => scaledPileInstance)
+        .forEach(scaledPileInstance => {
+          activePile.addChild(scaledPileInstance.graphics);
+        });
+
       renderRaf();
     }
 
@@ -1733,6 +1747,11 @@ const createPilingJs = (rootElement, initOptions = {}) => {
 
   const scaleBtnClick = (contextMenuElement, pileId) => () => {
     const pile = pileInstances.get(pileId);
+    if (pile.scale() > 1) {
+      store.dispatch(createAction.setScaledPiles([]));
+    } else {
+      store.dispatch(createAction.setScaledPiles([pileId]));
+    }
     pile.scaleToggle();
 
     hideContextMenu(contextMenuElement);
@@ -1813,7 +1832,14 @@ const createPilingJs = (rootElement, initOptions = {}) => {
         } else if (event.altKey) {
           results.forEach(result => {
             const pile = pileInstances.get(result.pileId);
-            if (pile.graphics.isHover) pile.animateScale();
+            if (pile.graphics.isHover) {
+              if (pile.scale() > 1) {
+                store.dispatch(createAction.setScaledPiles([]));
+              } else {
+                store.dispatch(createAction.setScaledPiles([result.pileId]));
+              }
+              pile.scaleToggle();
+            }
           });
         } else {
           results.forEach(result => {
@@ -1825,7 +1851,7 @@ const createPilingJs = (rootElement, initOptions = {}) => {
         }
       } else {
         store.dispatch(createAction.setFocusedPiles([]));
-        store.dispatch(createAction.setScaledPile([]));
+        store.dispatch(createAction.setScaledPiles([]));
       }
     }
   };
@@ -1882,11 +1908,8 @@ const createPilingJs = (rootElement, initOptions = {}) => {
     if (result.length !== 0) {
       if (event.altKey) {
         event.preventDefault();
-        store.dispatch(createAction.setScaledPile([result[0].pileId]));
-        const normalizedDeltaY = normalizeWheel(event).pixelY;
-        scalePile(result[0].pileId, normalizedDeltaY);
-        const graphics = pileInstances.get(result[0].pileId).graphics;
-        activePile.addChild(graphics);
+        store.dispatch(createAction.setScaledPiles([result[0].pileId]));
+        scalePile(result[0].pileId, normalizeWheel(event).pixelY);
       }
     }
   };
@@ -2009,8 +2032,8 @@ const createPilingJs = (rootElement, initOptions = {}) => {
         tempDepileBtn.innerHTML = 'close temp depile';
       }
 
-      if (pile.graphics.scale.x > 1.1) {
-        scaleBtn.innerHTML = 'scale down';
+      if (pile.scale() > 1) {
+        scaleBtn.innerHTML = 'Scale Down';
       }
 
       element.style.display = 'block';