+++ src/pile.js
@@ -364,7 +364,14 @@ const createPile = ({ initialItem, render, id, pubSub, store }) => {
 
   // Map to store calls for after the pile position animation
   const postPilePositionAnimation = new Map();
-  const animatePositionItems = (itemSprite, x, y, animator, isLastOne) => {
+  const animatePositionItems = (
+    itemSprite,
+    x,
+    y,
+    angle,
+    animator,
+    isLastOne
+  ) => {
     const targetScale = itemSprite.tmpTargetScale || itemSprite.scale.x;
     itemSprite.tmpTargetScale = undefined;
     delete itemSprite.tmpTargetScale;
@@ -372,15 +379,21 @@ const createPile = ({ initialItem, render, id, pubSub, store }) => {
     const tweener = createTweener({
       duration: 250,
       interpolator: interpolateVector,
-      endValue: [x, y, targetScale],
+      endValue: [x, y, targetScale, angle],
       getter: () => {
-        return [itemSprite.x, itemSprite.y, itemSprite.scale.x];
+        return [
+          itemSprite.x,
+          itemSprite.y,
+          itemSprite.scale.x,
+          itemSprite.angle
+        ];
       },
       setter: newValue => {
         itemSprite.x = newValue[0];
         itemSprite.y = newValue[1];
         itemSprite.scale.x = newValue[2];
         itemSprite.scale.y = itemSprite.scale.x;
+        itemSprite.angle = newValue[3];
       },
       onDone: () => {
         itemSprite.tmpTargetScale = undefined;
@@ -399,6 +412,7 @@ const createPile = ({ initialItem, render, id, pubSub, store }) => {
 
   const positionItems = (itemAlignment, itemRotation, animator, spacing) => {
     isPositioning = true;
+    let angle = 0;
     if (hasCover) {
       // matrix
       itemContainer.children.forEach((item, index) => {
@@ -410,11 +424,12 @@ const createPile = ({ initialItem, render, id, pubSub, store }) => {
           item,
           2 - spacing / 2,
           -padding,
+          angle,
           animator,
           index === itemContainer.children.length - 2
         );
       });
-    } else if (itemAlignment) {
+    } else if (itemAlignment || items.length === 1) {
       // image
       newItems.forEach(item => {
         const sprite = item.sprite;
@@ -477,11 +492,13 @@ const createPile = ({ initialItem, render, id, pubSub, store }) => {
           item,
           horizontalPadding + 2,
           verticalPadding + 2,
+          angle,
           animator,
           index === itemContainer.children.length - 1
         );
       });
     } else {
+      const { randomOffsetRange, randomRotationRange } = store.getState();
       let num = 0;
       newItems.forEach(item => {
         num++;
@@ -501,8 +518,8 @@ const createPile = ({ initialItem, render, id, pubSub, store }) => {
           delete item.tmpRelScale;
         }
 
-        let offsetX = getRandomArbitrary(-30, 30);
-        let offsetY = getRandomArbitrary(-30, 30);
+        let offsetX = getRandomArbitrary(...randomOffsetRange);
+        let offsetY = getRandomArbitrary(...randomOffsetRange);
 
         if (!Number.isNaN(+item.tmpAbsX) && !Number.isNaN(+item.tmpAbsY)) {
           offsetX += item.x;
@@ -517,17 +534,18 @@ const createPile = ({ initialItem, render, id, pubSub, store }) => {
           delete item.tmpAbsY;
         }
 
+        if (itemRotation) {
+          angle = getRandomArbitrary(...randomRotationRange);
+        }
+
         animatePositionItems(
           sprite,
           offsetX,
           offsetY,
+          angle,
           animator,
           num === newItems.size
         );
-
-        if (itemRotation) {
-          item.sprite.angle += getRandomArbitrary(-10, 10);
-        }
       });
     }
     newItems.clear();
@@ -586,10 +604,8 @@ const createPile = ({ initialItem, render, id, pubSub, store }) => {
     return oldScale !== newScale;
   };
 
-  let scaledUp = false;
   const scaleToggle = noAnimate => {
-    scale(scaledUp ? 1 : MAX_SCALE, noAnimate);
-    scaledUp = !scaledUp;
+    scale(getScale() > 1 ? 1 : MAX_SCALE, noAnimate);
   };
 
   const moveTo = (x, y) => {